cleaned up Sedov diagnostics
