+++ Diagnostics/Sedov/GNUmakefile
@@ -1,49 +1,36 @@
-NDEBUG := t
-MPI    :=
-OMP    :=
-MKVERBOSE :=t
+PRECISION = DOUBLE
+PROFILE = FALSE
+DEBUG = TRUE
+DIM = 3
 
-COMP := gfortran
+COMP = gnu
 
+USE_MPI = FALSE
+USE_OMP = FALSE
 
-# include the main Makefile stuff
-include $(AMREX_HOME)/Tools/F_mk/GMakedefs.mak
+USE_REACT = FALSE
 
-# core BoxLib directories
-BOXLIB_CORE := Src/F_BaseLib
+USE_ACC = FALSE
 
-# BoxLib stuff
-Fmpack += $(foreach dir, $(BOXLIB_CORE), $(AMREX_HOME)/$(dir)/GPackage.mak)
-Fmlocs += $(foreach dir, $(BOXLIB_CORE), $(AMREX_HOME)/$(dir))
+# programs to be compiled
+ALL: sedov_$(DIM)d.ex
 
+EOS_DIR := helmholtz
 
-# include the necessary GPackage.mak files that define this setup
-include $(Fmpack)
+NETWORK_DIR := aprox13
 
-# vpath defines the directories to search for the source files
-VPATH_LOCATIONS += $(Fmlocs)
+Bpack   := ./Make.package
+Blocs   := .
+# EXTERN_SEARCH = .
 
+CASTRO_HOME := ../..
 
+INCLUDE_LOCATIONS += $(AMREX_HOME)/Src/Extern/amrdata
+include $(AMREX_HOME)/Src/Extern/amrdata/Make.package
+vpathdir += $(AMREX_HOME)/Src/Extern/amrdata
 
-programs += fsedov1d
-#programs += fsedov2d_cart
-programs += fsedov2d_cyl_in_cartcoords
-# programs += fsedov2d_cyl_in_cylcoords
-#programs += fsedov2d_sph_in_cylcoords
-#programs += fsedov3d_cyl
-programs += fsedov3d_sph
-#programs += fsod2d
-#programs += fsod3d
+include $(CASTRO_HOME)/Exec/Make.Castro
 
-
-all: $(pnames)
-
-include $(AMREX_HOME)/Tools/F_mk/GMakerules.mak
-
-%.$(suf).exe:%.f90 $(objects)
-ifdef MKVERBOSE
-	$(LINK.f90) -o $@ $< $(objects) $(libraries)
-else
-	@echo "Linking $@ ... "
-	@$(LINK.f90) -o $@ $< $(objects) $(libraries)
-endif
+sedov_$(DIM)d.ex: $(objForExecs)
+	@echo Linking $@ ...
+	$(SILENT) $(PRELINK) $(CXX) $(CPPFLAGS) $(CXXFLAGS) $(LDFLAGS) -o $@ $^ $(libraries)
