+++ Diagnostics/Sedov/Sedov_F.H
@@ -36,7 +36,7 @@ void fextract2d_sph(const int* lo, const int* hi,
                     amrex::Real* dens_bin, amrex::Real* vel_bin,
                     amrex::Real* pres_bin, amrex::Real* e_bin,
                     int* volcount,
-                    int* imask, const int* mask_size, const int r1,
+                    int* imask, const int mask_size, const int r1,
                     const int dens_comp, const int xmom_comp,
                     const int ymom_comp,
                     const int pres_comp, const int rhoe_comp,
